fix: stop the request status page from crashing
the CI caught a bug where the page could try to show the task rank before the data had loaded, which would cause an error.

i added a check to make sure the rank data is actually there before we try to display it. this also updates the text in all languages to show the queue starting at 1.